# Memorify/src/main/java/com/moclase/Memorify/message/MessageRequest.java
package com.moclase.Memorify.message;


import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class MessageRequest {
    private String content;
    private String senderId;
    private String receiverId;
    private MessageType messageType;
    private String chatId;
}